pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "qlcplus"
version = "0.1.0"
description = "Python client library for controlling QLC+ lighting software via WebSocket API"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
authors = [
    {name = "Josh Lebedinsky"}
]
keywords = ["qlc+", "qlcplus", "dmx", "lighting", "websocket", "stage-lighting"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Graphics :: Presentation",
    "Topic :: System :: Hardware",
    "Typing :: Typed",
]
dependencies = [
    "websocket-client>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
]
osc = [
    "python-osc>=1.8.0",
]

[project.urls]
Repository = "https://github.com/joshlebed/qlc-config"

[project.scripts]
qlc = "ws_control:main"

[tool.hatch.build.targets.wheel]
packages = ["qlcplus"]

# Ruff configuration (linting + formatting)
[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "UP",     # pyupgrade
    "RUF",    # ruff-specific
]
ignore = [
    "E501",   # line too long (handled by formatter)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# Mypy configuration (type checking)
[tool.mypy]
python_version = "3.10"
strict = true
warn_return_any = true
warn_unused_ignores = true

[[tool.mypy.overrides]]
module = "websocket.*"
ignore_missing_imports = true

# Pytest configuration
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
